backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-rest-framework>=0.1.0",
    "django-stubs>=5.2.8",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.7",
    "dotenv>=0.9.9",
    "jwt>=1.4.0",
    "psycog>=0.0.1",
    "psycopg>=3.3.2",
    "psycopg2-binary>=2.9.11",
    "uuid>=1.30",
]
